refactor: unify markdown rendering functions
